instance_manager: Save PoI entity when a Deterministic Error happens
We want indexers to be able to provide the PoI for subgraphs that
failed for Deterministic reasons.

This commit remove an early return that forbid the entities to
be transacted to the database. Alongside that, there's a filter
to the entity modifications vector so that only the PoI gets
transacted/saved.