Fix prefetching bug when selecting by specific attributes
This fixes errors in prefetching phase for queries that orders the
selection set with the `orderBy` attribute.

The changes rely on complementing the selection set with the fields
used for ordering. The problematic part is that each visited selection
set is unaware of those ordering fields due to them not being in scope
during the recursive invocations of the `execute_selection_set`
function.

This led to malformed SQL queries that tried to `ORDER BY`
columns that were not being selected in sub-queries.

The solution presented here was to pass those complementary
fields as arguments along the recursive traversal and injecting them
into their respective selection sets.